+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- mongodb-memory-server can be omitted
+- api cache fix
+
 ## [1.2.5] - 2025-06-20
 
 ### Changed